# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/question/answer": {
            "post": {
                "description": "学生提交单个视频/片段内的题目答案，系统自动批改并返回结果",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "学生批量提交答题答案",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "学生答题参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/question.AnswerQuestionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/question.AnswerQuestionResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/question/generate/{video_id}": {
            "get": {
                "description": "根据视频ID自动生成课后练习题，供教师确认后发布",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "AI生成视频课后习题",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "视频ID参数",
                        "name": "video_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/question.GenerateVideoExerciseResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/stat/class": {
            "get": {
                "description": "获取班级在某个视频下的整体数据：正确率、完成率、薄弱知识点、高频暂停/回放片段",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistic"
                ],
                "summary": "获取班级整体学情统计",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "获取班级情况参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/statistic.GetClassStatReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/statistic.GetClassStatResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/stat/student/{video_id}": {
            "get": {
                "description": "获取学生单个视频的答题正确率、观看时长、暂停次数、薄弱知识点",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistic"
                ],
                "summary": "获取学生个人学情统计",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "视频ID",
                        "name": "video_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/statistic.GetStudentStatResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "question.AnswerQuestionReq": {
            "type": "object",
            "properties": {
                "studentanswers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/question.StudentAnswer"
                    }
                },
                "video_id": {
                    "type": "integer"
                }
            }
        },
        "question.AnswerQuestionResp": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/question.Result"
                    }
                }
            }
        },
        "question.GenerateVideoExerciseResp": {
            "type": "object",
            "properties": {
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/question.Question"
                    }
                }
            }
        },
        "question.KnowledgePoint": {
            "type": "object",
            "properties": {
                "knowledge_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "question.Question": {
            "type": "object",
            "properties": {
                "analysis": {
                    "type": "string"
                },
                "answer": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "knowledge_points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/question.KnowledgePoint"
                    }
                },
                "options": {
                    "type": "string"
                },
                "segment_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "question.Result": {
            "type": "object",
            "properties": {
                "analysis": {
                    "type": "string"
                },
                "answer": {
                    "type": "string"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "question_id": {
                    "type": "integer"
                }
            }
        },
        "question.StudentAnswer": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "question_id": {
                    "type": "integer"
                },
                "time_cost": {
                    "type": "integer"
                }
            }
        },
        "statistic.AverageKnowledgePoint": {
            "type": "object",
            "properties": {
                "average_master_score": {
                    "type": "number"
                },
                "knowledge_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "weak_rate": {
                    "description": "低于平均掌握程度分数的比例",
                    "type": "number"
                }
            }
        },
        "statistic.GetClassStatReq": {
            "type": "object",
            "properties": {
                "class_id": {
                    "type": "integer"
                },
                "video_id": {
                    "type": "integer"
                }
            }
        },
        "statistic.GetClassStatResp": {
            "type": "object",
            "properties": {
                "overview": {
                    "$ref": "#/definitions/statistic.Overview"
                },
                "top_pause_action": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/statistic.PauseAction"
                    }
                },
                "top_questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/statistic.QuestionStat"
                    }
                },
                "top_replay_action": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/statistic.ReplayAction"
                    }
                },
                "weak_knowledge_point": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/statistic.AverageKnowledgePoint"
                    }
                }
            }
        },
        "statistic.GetStudentStatResp": {
            "type": "object",
            "properties": {
                "correct_rate": {
                    "description": "题目正确率",
                    "type": "number"
                },
                "knowledge_points": {
                    "description": "薄弱知识点（薄弱知识点得分，要提前确定一个阈值）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/statistic.KnowledgePoint"
                    }
                },
                "pause_count": {
                    "description": "暂停总时长",
                    "type": "integer"
                },
                "status": {
                    "description": "任务完成情况：finished | in_progress | todo | expired",
                    "type": "string"
                },
                "time_cost": {
                    "description": "视频观看时长",
                    "type": "integer"
                }
            }
        },
        "statistic.KnowledgePoint": {
            "type": "object",
            "properties": {
                "knowledge_id": {
                    "type": "integer"
                },
                "master_score": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "statistic.Overview": {
            "type": "object",
            "properties": {
                "average_correct_rate": {
                    "description": "平均答题正确率",
                    "type": "number"
                },
                "average_time_cost": {
                    "description": "平均观看时长",
                    "type": "integer"
                },
                "complete_rate": {
                    "description": "任务完成率",
                    "type": "number"
                },
                "total_pause_count": {
                    "description": "总暂停时长",
                    "type": "integer"
                }
            }
        },
        "statistic.PauseAction": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "integer"
                },
                "pause_count": {
                    "type": "integer"
                },
                "segment_id": {
                    "type": "integer"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
        "statistic.QuestionStat": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "error_rate": {
                    "type": "number"
                },
                "question_id": {
                    "type": "integer"
                }
            }
        },
        "statistic.ReplayAction": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "integer"
                },
                "replay_count": {
                    "type": "integer"
                },
                "segment_id": {
                    "type": "integer"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
...
    }
}